+++ addon/components/paper-card-actions.js
@@ -1,0 +1,20 @@
+import Ember from 'ember';
+
+const { Component } = Ember;
+
+export default Component.extend({
+  tagName: 'md-card-actions',
+  classNameBindings: ['defaultClasses'],
+
+  didReceiveAttrs() {
+    this._super(...arguments);
+
+    // if the consumer didn't set layout classes, we should set them
+    // to the default (layout = row, layout align = end center)
+    let providedClasses = this.get('class');
+
+    if (!providedClasses || providedClasses.indexOf('layout-') === -1) {
+      this.set('defaultClasses', 'layout-row layout-align-end-center');
+    }
+  }
+});