+++ addon/components/paper-card-avatar.js
@@ -3,5 +3,5 @@ import Ember from 'ember';
 const { Component } = Ember;
 
 export default Component.extend({
-  tagName: 'md-card-content'
+  tagName: 'md-card-avatar'
 });