+++ addon/components/paper-card-title-media.js
@@ -3,5 +3,7 @@ import Ember from 'ember';
 const { Component } = Ember;
 
 export default Component.extend({
-  tagName: 'md-card-content'
+  tagName: 'md-card-title-media',
+
+  size: 'md'
 });