+++ app/components/paper-card-avatar.js
@@ -1,0 +1,3 @@
+import PaperCardAvatar from 'ember-paper/components/paper-card-avatar';
+
+export default PaperCardAvatar;