+++ app/components/paper-card-header-content.js
@@ -1,0 +1,3 @@
+import PaperCardHeaderContent from 'ember-paper/components/paper-card-header-content';
+
+export default PaperCardHeaderContent;